requirements.txt
amqp==5.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.6.1
Automat==20.2.0
billiard==3.6.4.0
boto3==1.24.42
botocore==1.27.42
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Deprecated==1.2.13
Django==4.0.6
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-redis==5.2.0
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
jmespath==1.0.1
kombu==5.2.4
mutagen==1.45.1
packaging==21.3
Pillow==9.2.0
pip==22.2.1
pluggy==1.0.0
prompt-toolkit==3.0.30
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
redis==4.3.4
requests==2.28.1
s3transfer==0.6.0
service-identity==21.1.0
setuptools==62.1.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wheel==0.37.1
whitenoise==6.2.0
wrapt==1.14.1
zope.interface==5.4.0
